Added basic support for generating API documentation using doxygen
If cmake is run and doxygen is found, add a make target called "doc",
which builds API documention in doc/html. This target is not included
in the default "all" target, you have to explicitely run "make doc" to
generate the documentation. Point your browser to doc/html/index.html
to browse the generated documentation.

Documentation is generated for source code in the directories src/ and
src/util/, but not for src/lua/, src/jthread/ and src/sqlite/.

If graphviz is installed, in particular, if the "dot" binary is found,
doxygen is configured to generate various kinds of diagrams. Note that
due to this, the first run of doxygen can take a while.